requirements.txt
Pillow 
rembg 
opencv-python 
PyMatting
streamlit
moviepy